pyproject.toml
[project]
name = "codegen-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chroma>=0.2.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "langsmith>=0.7.30",
    "pypdf>=6.10.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["codegen_agent"]